+++ tests/integration/resources/models/mvp_model/__init__.py
@@ -1,1 +1,0 @@
-"""A Model of the MVP."""